refactor: move components/ui to components/Utilities
Consolidate UI components into the Utilities directory for better
organization and consistency with the feature-focused structure.

Changes:
- Move Spinner.tsx from components/ui to components/Utilities
- Move StatsCard.tsx from components/ui to components/Utilities
- Update imports in app pages (page.tsx, stats/page.tsx, books/[id]/page.tsx)
- Update import in components/Modals/BaseModal.tsx
- Remove empty components/ui directory

All 2118 tests pass.
